requirementsbak.txt
# Basic dependencies
numpy>=1.21.0
torch>=1.9.0
torchaudio>=0.9.0
scipy>=1.7.0
sounddevice>=0.4.5
soundfile>=0.10.3
psutil>=5.9.8
pydub>=0.25.1
librosa>=0.9.0
noisereduce>=3.0.3
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0

librosa>=0.9.2      # 音频增强
scipy>=1.7.0        # 音频处理
resampy>=0.4.2      # 重采样



# ASR dependencies
faster-whisper>=0.9.0
vosk>=0.3.45
ffmpeg-python>=0.2.0

# Translation dependencies
transformers>=4.30.0
sentencepiece>=0.1.99
deep-translator>=1.11.4
langdetect>=1.0.9
sacremoses>=0.1.1

# TTS dependencies
edge-tts>=6.1.9
TTS>=0.22.0
bark>=0.0.1a0
resemblyzer>=0.1.1
pynvml>=12.0.0
f5-tts==1.1.4
cached_path>=1.1.6

# RVC voice cloning
fairseq==0.12.2
omegaconf==2.0.6
hydra-core==1.0.7
antlr4-python3-runtime==4.8
pyyaml>=6.0
torchcrepe>=0.0.12
resampy>=0.4.2
tensorboard>=2.13.0
tensorboardX>=2.6
matplotlib>=3.7.1
scikit-learn>=1.3.0
scikit-image>=0.21.0
pandas>=1.4.0,<2.0.0

# Audio processing
webrtcvad>=2.0.10

# Utilities
tqdm>=4.65.0
pillow>=10.0.0

# Core dependencies
fastapi>=0.115.12
uvicorn>=0.34.0
python-multipart>=0.0.20
starlette>=0.46.1
pydantic>=2.10.6
typer>=0.15.2

# UI
gradio>=5.23.0
rich>=13.9.4

# Network testing
speedtest-cli>=2.1.3

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

# Additional dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Translation dependencies
googletrans>=3.1.0a0
translate>=3.6.1

accelerate>=0.26.0

requests tqdm

torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118 --no-cache-dir